pyproject.toml
[project]
name = "sentinel-core-mvp"
version = "0.1.0"
description = "AI-powered code analysis and automated bug detection/repair system using Tree-sitter, NetworkX CPG, and multi-agent architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "langchain>=1.2.15",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.7",
    "networkx>=3.4",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "black>=24.0",
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel_core"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
